requirements.txt
python-dotenv==1.0.1
requests==2.32.3
pytz==2024.2
gspread==6.1.4
oauth2client==4.1.3
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
